requirements.txt
numpy==1.23.0
Pillow==9.4.0
Pillow==10.1.0
streamlit==1.24.0
streamlit==1.24.1
tensorflow==2.10.0
